TIP-843: Move processor to enrichment
