+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/processors.yml
@@ -1,0 +1,101 @@
+services:
+
+    pim_connector.processor.denormalization.product:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductProcessor'
+        arguments:
+            - '@pim_catalog.repository.product'
+            - '@pim_connector.processor.denormalization.product_to_import'
+            - '@pim_catalog.entity_with_family_variant.add_parent_to_product'
+            - '@pim_catalog.updater.product'
+            - '@pim_catalog.validator.product'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+            - '@pim_catalog.comparator.filter.product'
+            - '@pim_connector.processor.attribute_filter.product'
+
+    pim_connector.processor.denormalization.product_to_import:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\FindProductToImport'
+        arguments:
+            - '@pim_catalog.repository.product'
+            - '@pim_catalog.builder.product'
+
+    pim_connector.processor.denormalization.root_product_model:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductModelProcessor'
+        arguments:
+            - '@pim_catalog.factory.product_model'
+            - '@pim_catalog.updater.product_model'
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.validator.product'
+            - '@pim_catalog.comparator.filter.product_model'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+            - '@pim_connector.processor.attribute_filter.product_model'
+            - 'root_product_model'
+
+    pim_connector.processor.denormalization.sub_product_model:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductModelProcessor'
+        arguments:
+            - '@pim_catalog.factory.product_model'
+            - '@pim_catalog.updater.product_model'
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.validator.product'
+            - '@pim_catalog.comparator.filter.product_model'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+            - '@pim_connector.processor.attribute_filter.product_model'
+            - 'sub_product_model'
+
+    pim_connector.processor.denormalization.product_model_loader:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductModelLoaderProcessor'
+        arguments:
+            - '@pim_catalog.repository.product_model'
+
+    pim_connector.processor.denormalization.product_association:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductAssociationProcessor'
+        arguments:
+            - '@pim_catalog.repository.product'
+            - '@pim_catalog.updater.product'
+            - '@pim_catalog.validator.product'
+            - '@pim_catalog.comparator.filter.product_association'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+
+    pim_connector.processor.denormalization.product_model_association:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductModelAssociationProcessor'
+        arguments:
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.updater.product_model'
+            - '@pim_catalog.validator.product_model'
+            - '@pim_catalog.comparator.filter.product_association'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+
+    pim_connector.processor.normalization.product:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Normalizer\ProductProcessor'
+        arguments:
+            - '@pim_catalog.normalizer.standard.product'
+            - '@pim_catalog.repository.channel'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_connector.processor.bulk_media_fetcher'
+            - '@pim_catalog.values_filler.product'
+            - '@pim_connector.doctrine.cache_clearer'
+
+    pim_connector.processor.normalization.product_model:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Normalizer\ProductProcessor'
+        arguments:
+            - '@pim_catalog.normalizer.standard.product_model'
+            - '@pim_catalog.repository.channel'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_connector.processor.bulk_media_fetcher'
+            - '@pim_catalog.values_filler.entity_with_family_variant'
+            - '@pim_connector.doctrine.cache_clearer'
+
+    pim_connector.processor.attribute_filter.product_model:
+        class: 'Pim\Component\Catalog\ProductModel\Filter\ProductModelAttributeFilter'
+        arguments:
+            - '@pim_catalog.repository.family_variant'
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.repository.cached_attribute'
+
+    pim_connector.processor.attribute_filter.product:
+        class: 'Pim\Component\Catalog\ProductModel\Filter\ProductAttributeFilter'
+        arguments:
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.repository.cached_family'
+            - '@pim_catalog.repository.product'
+            - '@pim_catalog.repository.cached_attribute'
