+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/Denormalizer/ProductAssociationProcessor.php
@@ -1,9 +1,8 @@
 <?php
-declare(strict_types=1);
 
-namespace Pim\Component\Connector\Processor\Denormalization;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Tool\Component\Batch\Item\ItemProcessorInterface;
 use Akeneo\Tool\Component\Batch\Step\StepExecutionAwareInterface;
 use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
@@ -11,20 +10,20 @@ use Akeneo\Tool\Component\StorageUtils\Exception\PropertyException;
 use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Tool\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Pim\Component\Catalog\Comparator\Filter\FilterInterface;
-use Pim\Component\Catalog\Model\EntityWithAssociationsInterface;
+use Pim\Component\Connector\Processor\Denormalization\AbstractProcessor;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Exception\InvalidArgumentException;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
- * @author    Adrien Pétremann <adrien.petremann@akeneo.com>
- * @copyright 2018 Akeneo SAS (https://www.akeneo.com)
+ * Product association import processor
+ *
+ * @author    Julien Sanchez <julien@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductModelAssociationProcessor extends AbstractProcessor implements
-    ItemProcessorInterface,
-    StepExecutionAwareInterface
+class ProductAssociationProcessor extends AbstractProcessor implements ItemProcessorInterface, StepExecutionAwareInterface
 {
     /** @var IdentifiableObjectRepositoryInterface */
     protected $repository;
@@ -39,17 +38,17 @@ class ProductModelAssociationProcessor extends AbstractProcessor implements
     protected $detacher;
 
     /** @var FilterInterface */
-    protected $associationsFilter;
+    protected $productAssocFilter;
 
     /** @var bool */
     protected $enabledComparison = true;
 
     /**
-     * @param IdentifiableObjectRepositoryInterface $repository
-     * @param ObjectUpdaterInterface                $updater
-     * @param ValidatorInterface                    $validator
-     * @param FilterInterface                       $productAssocFilter
-     * @param ObjectDetacherInterface               $detacher
+     * @param IdentifiableObjectRepositoryInterface $repository         product repository
+     * @param ObjectUpdaterInterface                $updater            product updater
+     * @param ValidatorInterface                    $validator          validator of the object
+     * @param FilterInterface                       $productAssocFilter product association filter
+     * @param ObjectDetacherInterface               $detacher           detacher to remove it from UOW when skip
      */
     public function __construct(
         IdentifiableObjectRepositoryInterface $repository,
@@ -63,7 +62,7 @@ class ProductModelAssociationProcessor extends AbstractProcessor implements
         $this->repository = $repository;
         $this->updater = $updater;
         $this->validator = $validator;
-        $this->associationsFilter = $productAssocFilter;
+        $this->productAssocFilter = $productAssocFilter;
         $this->detacher = $detacher;
     }
 
@@ -77,98 +76,94 @@ class ProductModelAssociationProcessor extends AbstractProcessor implements
             $item
         );
 
-        if (!isset($item['code'])) {
-            $this->skipItemWithMessage($item, 'The code must be filled');
+        if (!isset($item['identifier'])) {
+            $this->skipItemWithMessage($item, 'The identifier must be filled');
         }
 
-        $entity = $this->findEntity($item['code'], $item);
-        if (null === $entity) {
-            $this->skipItemWithMessage($item, sprintf('No product model with code "%s" has been found', $item['code']));
+        $product = $this->findProduct($item['identifier'], $item);
+        if (null === $product) {
+            $this->skipItemWithMessage($item, sprintf('No product with identifier "%s" has been found', $item['identifier']));
         }
 
         $parameters = $this->stepExecution->getJobParameters();
         $enabledComparison = $parameters->get('enabledComparison');
         if ($enabledComparison) {
-            $item = $this->filterIdenticalData($entity, $item);
+            $item = $this->filterIdenticalData($product, $item);
 
             if (empty($item)) {
-                $this->detach($entity);
-                $this->stepExecution->incrementSummaryInfo('product_model_skipped_no_diff');
+                $this->detachProduct($product);
+                $this->stepExecution->incrementSummaryInfo('product_skipped_no_diff');
 
                 return null;
             }
         } elseif (!$this->hasImportedAssociations($item)) {
-            $this->detach($entity);
-            $this->stepExecution->incrementSummaryInfo('product_model_skipped_no_associations');
+            $this->detachProduct($product);
+            $this->stepExecution->incrementSummaryInfo('product_skipped_no_associations');
 
             return null;
         }
 
         try {
-            $this->update($entity, $item);
+            $this->updateProduct($product, $item);
         } catch (PropertyException | InvalidArgumentException | AccessDeniedException $exception) {
-            $this->detach($entity);
+            $this->detachProduct($product);
             $this->skipItemWithMessage($item, $exception->getMessage(), $exception);
         }
 
-        $violations = $this->validateAssociations($entity);
+        $violations = $this->validateProductAssociations($product);
         if ($violations && $violations->count() > 0) {
-            $this->detach($entity);
+            $this->detachProduct($product);
             $this->skipItemWithConstraintViolations($item, $violations);
         }
 
-        return $entity;
+        return $product;
     }
 
     /**
-     * @param ProductModelInterface $product
-     * @param array                 $item
+     * @param ProductInterface $product
+     * @param array            $item
      *
      * @return array
      */
-    protected function filterIdenticalData(ProductModelInterface $product, array $item): array
+    protected function filterIdenticalData(ProductInterface $product, array $item)
     {
-        return $this->associationsFilter->filter($product, $item);
+        return $this->productAssocFilter->filter($product, $item);
     }
 
     /**
-     * @param ProductModelInterface $productModel
-     * @param array                 $item
+     * @param ProductInterface $product
+     * @param array            $item
      *
      * @throws PropertyException
      */
-    protected function update(ProductModelInterface $productModel, array $item): void
+    protected function updateProduct(ProductInterface $product, array $item)
     {
-        $this->updater->update($productModel, $item);
+        $this->updater->update($product, $item);
     }
 
     /**
      * @param string $identifier
      * @param array  $item
      *
-     * @return null|ProductModelInterface
-     *
-     * @throws \Akeneo\Component\Batch\Item\InvalidItemException
+     * @return null|ProductInterface
      */
-    public function findEntity(string $identifier, array $item): ?ProductModelInterface
+    public function findProduct(string $identifier, array $item): ?ProductInterface
     {
         try {
             return $this->repository->findOneByIdentifier($identifier);
         } catch (AccessDeniedException $e) {
             $this->skipItemWithMessage($item, $e->getMessage(), $e);
         }
-
-        return null;
     }
 
     /**
-     * @param EntityWithAssociationsInterface $product
+     * @param ProductInterface $product
      *
      * @throws \InvalidArgumentException
      *
      * @return ConstraintViolationListInterface|null
      */
-    protected function validateAssociations(EntityWithAssociationsInterface $product): ?ConstraintViolationListInterface
+    protected function validateProductAssociations(ProductInterface $product)
     {
         $associations = $product->getAssociations();
         foreach ($associations as $association) {
@@ -182,15 +177,15 @@ class ProductModelAssociationProcessor extends AbstractProcessor implements
     }
 
     /**
-     * Detaches the product model from the unit of work is the responsibility of the writer but in this case we
+     * Detaches the product from the unit of work is the responsibility of the writer but in this case we
      * want ensure that an updated and invalid product will not be used in the association processor.
-     * Also we don't want to keep skipped product models in memory
+     * Also we don't want to keep skipped products in memory
      *
-     * @param EntityWithAssociationsInterface $productModel
+     * @param ProductInterface $product
      */
-    protected function detach(EntityWithAssociationsInterface $productModel): void
+    protected function detachProduct(ProductInterface $product)
     {
-        $this->detacher->detach($productModel);
+        $this->detacher->detach($product);
     }
 
     /**
@@ -200,18 +195,14 @@ class ProductModelAssociationProcessor extends AbstractProcessor implements
      *
      * @return bool
      */
-    protected function hasImportedAssociations(array $item): bool
+    protected function hasImportedAssociations(array $item)
     {
         if (!isset($item['associations'])) {
             return false;
         }
 
         foreach ($item['associations'] as $association) {
-            $hasProductAssoc = !empty($association['products']);
-            $hasGroupAssoc = !empty($association['groups']);
-            $hasProductModelAssoc = !empty($association['product_models']);
-
-            if ($hasProductAssoc || $hasGroupAssoc || $hasProductModelAssoc) {
+            if (!empty($association['products']) || !empty($association['groups'])) {
                 return true;
             }
         }
