+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/Denormalizer/ProductModelAssociationProcessor.php
@@ -1,8 +1,9 @@
 <?php
+declare(strict_types=1);
 
-namespace Pim\Component\Connector\Processor\Denormalization;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Tool\Component\Batch\Item\ItemProcessorInterface;
 use Akeneo\Tool\Component\Batch\Step\StepExecutionAwareInterface;
 use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
@@ -10,19 +11,21 @@ use Akeneo\Tool\Component\StorageUtils\Exception\PropertyException;
 use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Tool\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Pim\Component\Catalog\Comparator\Filter\FilterInterface;
+use Pim\Component\Catalog\Model\EntityWithAssociationsInterface;
+use Pim\Component\Connector\Processor\Denormalization\AbstractProcessor;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Exception\InvalidArgumentException;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
- * Product association import processor
- *
- * @author    Julien Sanchez <julien@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @author    Adrien Pétremann <adrien.petremann@akeneo.com>
+ * @copyright 2018 Akeneo SAS (https://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductAssociationProcessor extends AbstractProcessor implements ItemProcessorInterface, StepExecutionAwareInterface
+class ProductModelAssociationProcessor extends AbstractProcessor implements
+    ItemProcessorInterface,
+    StepExecutionAwareInterface
 {
     /** @var IdentifiableObjectRepositoryInterface */
     protected $repository;
@@ -37,17 +40,17 @@ class ProductAssociationProcessor extends AbstractProcessor implements ItemProce
     protected $detacher;
 
     /** @var FilterInterface */
-    protected $productAssocFilter;
+    protected $associationsFilter;
 
     /** @var bool */
     protected $enabledComparison = true;
 
     /**
-     * @param IdentifiableObjectRepositoryInterface $repository         product repository
-     * @param ObjectUpdaterInterface                $updater            product updater
-     * @param ValidatorInterface                    $validator          validator of the object
-     * @param FilterInterface                       $productAssocFilter product association filter
-     * @param ObjectDetacherInterface               $detacher           detacher to remove it from UOW when skip
+     * @param IdentifiableObjectRepositoryInterface $repository
+     * @param ObjectUpdaterInterface                $updater
+     * @param ValidatorInterface                    $validator
+     * @param FilterInterface                       $productAssocFilter
+     * @param ObjectDetacherInterface               $detacher
      */
     public function __construct(
         IdentifiableObjectRepositoryInterface $repository,
@@ -61,7 +64,7 @@ class ProductAssociationProcessor extends AbstractProcessor implements ItemProce
         $this->repository = $repository;
         $this->updater = $updater;
         $this->validator = $validator;
-        $this->productAssocFilter = $productAssocFilter;
+        $this->associationsFilter = $productAssocFilter;
         $this->detacher = $detacher;
     }
 
@@ -75,94 +78,98 @@ class ProductAssociationProcessor extends AbstractProcessor implements ItemProce
             $item
         );
 
-        if (!isset($item['identifier'])) {
-            $this->skipItemWithMessage($item, 'The identifier must be filled');
+        if (!isset($item['code'])) {
+            $this->skipItemWithMessage($item, 'The code must be filled');
         }
 
-        $product = $this->findProduct($item['identifier'], $item);
-        if (null === $product) {
-            $this->skipItemWithMessage($item, sprintf('No product with identifier "%s" has been found', $item['identifier']));
+        $entity = $this->findEntity($item['code'], $item);
+        if (null === $entity) {
+            $this->skipItemWithMessage($item, sprintf('No product model with code "%s" has been found', $item['code']));
         }
 
         $parameters = $this->stepExecution->getJobParameters();
         $enabledComparison = $parameters->get('enabledComparison');
         if ($enabledComparison) {
-            $item = $this->filterIdenticalData($product, $item);
+            $item = $this->filterIdenticalData($entity, $item);
 
             if (empty($item)) {
-                $this->detachProduct($product);
-                $this->stepExecution->incrementSummaryInfo('product_skipped_no_diff');
+                $this->detach($entity);
+                $this->stepExecution->incrementSummaryInfo('product_model_skipped_no_diff');
 
                 return null;
             }
         } elseif (!$this->hasImportedAssociations($item)) {
-            $this->detachProduct($product);
-            $this->stepExecution->incrementSummaryInfo('product_skipped_no_associations');
+            $this->detach($entity);
+            $this->stepExecution->incrementSummaryInfo('product_model_skipped_no_associations');
 
             return null;
         }
 
         try {
-            $this->updateProduct($product, $item);
+            $this->update($entity, $item);
         } catch (PropertyException | InvalidArgumentException | AccessDeniedException $exception) {
-            $this->detachProduct($product);
+            $this->detach($entity);
             $this->skipItemWithMessage($item, $exception->getMessage(), $exception);
         }
 
-        $violations = $this->validateProductAssociations($product);
+        $violations = $this->validateAssociations($entity);
         if ($violations && $violations->count() > 0) {
-            $this->detachProduct($product);
+            $this->detach($entity);
             $this->skipItemWithConstraintViolations($item, $violations);
         }
 
-        return $product;
+        return $entity;
     }
 
     /**
-     * @param ProductInterface $product
-     * @param array            $item
+     * @param ProductModelInterface $product
+     * @param array                 $item
      *
      * @return array
      */
-    protected function filterIdenticalData(ProductInterface $product, array $item)
+    protected function filterIdenticalData(ProductModelInterface $product, array $item): array
     {
-        return $this->productAssocFilter->filter($product, $item);
+        return $this->associationsFilter->filter($product, $item);
     }
 
     /**
-     * @param ProductInterface $product
-     * @param array            $item
+     * @param ProductModelInterface $productModel
+     * @param array                 $item
      *
      * @throws PropertyException
      */
-    protected function updateProduct(ProductInterface $product, array $item)
+    protected function update(ProductModelInterface $productModel, array $item): void
     {
-        $this->updater->update($product, $item);
+        $this->updater->update($productModel, $item);
     }
 
     /**
      * @param string $identifier
      * @param array  $item
      *
-     * @return null|ProductInterface
+     * @return null|ProductModelInterface
+     *
+     * @throws \Akeneo\Component\Batch\Item\InvalidItemException
      */
-    public function findProduct(string $identifier, array $item): ?ProductInterface
+    public function findEntity(string $identifier, array $item): ?ProductModelInterface
     {
         try {
             return $this->repository->findOneByIdentifier($identifier);
         } catch (AccessDeniedException $e) {
             $this->skipItemWithMessage($item, $e->getMessage(), $e);
         }
+
+        return null;
     }
 
     /**
-     * @param ProductInterface $product
+     * @param EntityWithAssociationsInterface $product
      *
      * @throws \InvalidArgumentException
      *
      * @return ConstraintViolationListInterface|null
      */
-    protected function validateProductAssociations(ProductInterface $product)
+    protected function validateAssociations(EntityWithAssociationsInterface $product): ?ConstraintViolationListInterface
     {
         $associations = $product->getAssociations();
         foreach ($associations as $association) {
@@ -176,15 +183,15 @@ class ProductAssociationProcessor extends AbstractProcessor implements ItemProce
     }
 
     /**
-     * Detaches the product from the unit of work is the responsibility of the writer but in this case we
+     * Detaches the product model from the unit of work is the responsibility of the writer but in this case we
      * want ensure that an updated and invalid product will not be used in the association processor.
-     * Also we don't want to keep skipped products in memory
+     * Also we don't want to keep skipped product models in memory
      *
-     * @param ProductInterface $product
+     * @param EntityWithAssociationsInterface $productModel
      */
-    protected function detachProduct(ProductInterface $product)
+    protected function detach(EntityWithAssociationsInterface $productModel): void
     {
-        $this->detacher->detach($product);
+        $this->detacher->detach($productModel);
     }
 
     /**
@@ -194,14 +201,18 @@ class ProductAssociationProcessor extends AbstractProcessor implements ItemProce
      *
      * @return bool
      */
-    protected function hasImportedAssociations(array $item)
+    protected function hasImportedAssociations(array $item): bool
     {
         if (!isset($item['associations'])) {
             return false;
         }
 
         foreach ($item['associations'] as $association) {
-            if (!empty($association['products']) || !empty($association['groups'])) {
+            $hasProductAssoc = !empty($association['products']);
+            $hasGroupAssoc = !empty($association['groups']);
+            $hasProductModelAssoc = !empty($association['product_models']);
+
+            if ($hasProductAssoc || $hasGroupAssoc || $hasProductModelAssoc) {
                 return true;
             }
         }
