+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Processor/Denormalizer/ProductModelLoaderProcessorSpec.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Processor\Denormalization;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer;
 
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Pim\Component\Catalog\Repository\ProductModelRepositoryInterface;
-use Pim\Component\Connector\Processor\Denormalization\ProductModelLoaderProcessor;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Denormalizer\ProductModelLoaderProcessor;
 use PhpSpec\ObjectBehavior;
 
 class ProductModelLoaderProcessorSpec extends ObjectBehavior
@@ -16,10 +16,11 @@ class ProductModelLoaderProcessorSpec extends ObjectBehavior
         $this->setStepExecution($stepExecution);
         $repository->getIdentifierProperties()->willReturn(['code']);
     }
-    
+
     function it_is_initializable()
     {
-        $this->shouldHaveType(ProductModelLoaderProcessor::class);
+        $this->shouldHaveType(
+           ProductModelLoaderProcessor::class);
     }
 
     function it_load_a_product_model($repository, ProductModelInterface $productModel)
